% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAFTA-package.R
\docType{package}
\name{CAFTA-package}
\alias{CAFTA}
\alias{CAFTA-package}
\title{CAFTA: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/signorinoy/CAFTA}
  \item \url{http://gongziyang.com/CAFTA/}
  \item Report bugs at \url{https://github.com/signorinoy/CAFTA/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
