% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cafta.R
\name{coef.cafta}
\alias{coef.cafta}
\title{Extract Model Coefficients from a cafta Object}
\usage{
\method{coef}{cafta}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cafta}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A numeric vector containing the estimated coefficients.
}
\description{
This method extracts the estimated coefficients from an object
of class \code{cafta}.
}
\seealso{
\code{\link{cafta}}
}
