% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cafta.R
\name{update.cafta}
\alias{update.cafta}
\title{Update a CAFTA Model Object}
\usage{
\method{update}{cafta}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class \code{cafta}.}

\item{newdata}{A data frame containing the new data to be used for updating
the model. The data frame must contain the same variables as those used in
the original model.}

\item{...}{Additional arguments (currently unused).}
}
\value{
An updated object of class \code{cafta}.
}
\description{
Updates a fitted CAFTA model object with new data.
}
\seealso{
\code{\link{cafta}}, \code{\link[stats]{update}}
}
