% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cafta.R
\name{vcov.cafta}
\alias{vcov.cafta}
\title{Calculate Variance-Covariance Matrix for cafta Objects}
\usage{
\method{vcov}{cafta}(object, ...)
}
\arguments{
\item{object}{An object of class \code{cafta}.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A variance-covariance matrix.
}
\description{
Computes the variance-covariance matrix for objects of
class \code{cafta}. If the \code{robust} slot of the object is \code{TRUE},
a robust variance-covariance matrix is calculated using the sandwich
estimator. Otherwise, the standard variance-covariance matrix is computed
as the inverse of the Hessian matrix.
}
\seealso{
\code{\link{cafta}}, \code{\link[stats]{vcov}}
}
